refactor: RacingGameResult controller패키지로 이동
