+++ src/main/java/racing/RacingGameStart.java
@@ -2,7 +2,7 @@ package racing;
 
 import racing.domain.car.RacingCars;
 import racing.domain.game.RacingGame;
-import racing.domain.game.RacingGameResult;
+import racing.controller.RacingGameResult;
 import racing.view.InputView;
 import racing.view.ResultView;
